Unify search APIs
